refactor; Created command package
