+++ cognite_toolkit/_cdf_tk/commands/__init__.py
@@ -1,0 +1,4 @@
+from .build import BuildCommand
+from .clean import CleanCommand
+
+__all__ = ["BuildCommand", "CleanCommand"]
